build: let old build system coexist with waf
Rename configure and the makefile, because otherwise it will severely
confuse users. Normally ./configure && make is the way to build, which
in our case would use the old build system.

./old-configure will actually create the Makefile, so "make" will still
work after ./old-configure is run.

Don't create config.h in the top-level directory, because the waf build
would pick it up and create a major confusing mess. Create it in
old_build/ instead, and add it as include path to the CFLAGS. config.mak
could stay in the root, but move it to old_build too for symmetry.