build: verify downloaded waf binary
I don't feel particularly good about executing automatically downloaded
files from the internet without verification. Verify the downloaded
executable with sha256sum, which is part of GNU coreutils, and should be
ubiquitous enough.

Now it should be impossible to run a downloaded file other than what the
hash stored in autogen.sh indicates, which in turn is bound to the git
commit hash.

If she256sum is not available, print a warning, and don't make waf
executable.